be ref(): fix module tight coupling
